+++ lib/flutter_water_flow.dart
@@ -4,6 +4,7 @@ library flutter_water_flow;
 export 'src/water_flow_view.dart';
 
 // 数据模型
+export 'src/models/water_flow_base_item.dart';
 export 'src/models/water_flow_item.dart';
 export 'src/models/info_flow_card_model.dart';