+++ lib/src/models/water_flow_base_item.dart
@@ -1,0 +1,5 @@
+import 'package:flutter_water_flow/flutter_water_flow.dart';
+
+abstract class WaterFlowBaseItem {
+  WaterFlowItemType get type;
+}