+++ lib/src/models/water_flow_item.dart
@@ -1,9 +1,12 @@
+import 'water_flow_base_item.dart';
+
 enum WaterFlowItemType {
   normal,
   ad,
 }
 
-class WaterFlowItem {
+class WaterFlowItem implements WaterFlowBaseItem {
+  @override
   final WaterFlowItemType type;
   final int index;
   final dynamic data;